mcp-server/requirements.txt
# MCP SDK
mcp>=1.0.0

# Web framework for HTTP transport
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Additional utilities
python-multipart>=0.0.6
pydantic>=2.0.0

# Authentication & Authorization
pyjwt[crypto]>=2.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
aiosqlite>=0.19.0

# Rate Limiting
slowapi>=0.1.9

# Google Cloud Services
google-cloud-secret-manager>=2.16.0
google-cloud-logging>=3.5.0

# Security
python-dotenv>=1.0.0
